+++ source/renderer/app/themes/daedalus/incentivized-testnet.js
@@ -619,13 +619,20 @@ export const INCENTIVIZED_TESTNET_THEME_OUTPUT = {
     '--theme-splash-network-background-color7': 'rgba(18, 19, 38, 0.91)',
     '--theme-splash-network-background-color8': 'rgba(18, 19, 38, 1)',
     '--theme-splash-network-logo-fill-color': '#fff',
-    '--theme-splash-network-title-color': '#fafbfc',
+    '--theme-splash-network-overlay-button-background-color':
+      'rgba(255, 255, 255, 0.1)',
+    '--theme-splash-network-overlay-button-background-color-hover': '#fff',
+    '--theme-splash-network-overlay-button-border-color': '#fff',
+    '--theme-splash-network-overlay-button-text-color': '#fff',
+    '--theme-splash-network-overlay-button-text-color-hover': '#2a2b3c',
+    '--theme-splash-network-title-color': '#fff',
     '--theme-splash-network-subTitle1-color': '#eb2256',
     '--theme-splash-network-subTitle2-color': '#eb2256',
     '--theme-splash-network-description-background-color':
       'rgba(255, 255, 255, 0.1)',
     '--theme-splash-network-description-color': 'rgba(255, 255, 255, 0.7)',
-    '--theme-splash-network-learn-more-color': '#fafbfc',
+    '--theme-splash-network-description-bold-color': '#fff',
+    '--theme-splash-network-learn-more-color': '#fff',
   },
   stakePools: {
     '--theme-staking-stake-pools-title-color': '#ffffff',