+++ storybook/stories/staking/StakePools.stories.js
@@ -6,9 +6,7 @@ import { action } from '@storybook/addon-actions';
 import StakePools from '../../../source/renderer/app/components/staking/stake-pools/StakePools';
 import STAKE_POOLS from '../../../source/renderer/app/config/stakingStakePools.dummy.json';
 
-const currentTheme = sessionStorage.getItem('themeName') || 'light-blue';
-
-export const StakePoolsStory = () => (
+export const StakePoolsStory = (props: { currentTheme: string }) => (
   <StakePools
     stakePoolsList={STAKE_POOLS.slice(
       0,
@@ -26,7 +24,7 @@ export const StakePoolsStory = () => (
       STAKE_POOLS[36],
     ]}
     onOpenExternalLink={() => {}}
-    currentTheme={currentTheme}
+    currentTheme={props.currentTheme}
     onDelegate={action('onDelegate')}
   />
 );